scaffolder/templates/django/requirements.txt
Django==1.10
django-autofixture==0.12.1{% if deployable_in_heroku %}
dj-database-url==0.4.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
{% endif %}
